Login

+++ Home/views.py
@@ -4,6 +4,9 @@ from django.http import HttpResponse
 from .models import Student
 from .forms import StudentCreationForm, UpdateStudentForm
 
+from django.contrib.auth.decorators import login_required
+
+@login_required(login_url='login')
 def index (request):
 
     """
@@ -35,8 +38,10 @@ def retrive_student(request,id):
 #         return HttpResponse("Student Added Successfully")
 #     else :
 #         return render(request,'Home/add_student.html')
-    
+@login_required(login_url='login')   
 def add_student (request):
+    if not request.user.is_superuser:
+        return redirect('index')
     form = StudentCreationForm()
     if request.method == 'POST':
         form = StudentCreationForm(request.POST, request.FILES)
@@ -70,8 +75,10 @@ def add_student (request):
 #     else :
 #         return render(request,'Home/add_student.html',{'student':student})
 
-
+@login_required(login_url='login')
 def update_student(request,id):
+    if not request.user.is_superuser:
+        return redirect('index')
     student = Student.objects.get(pk=id)
     form = UpdateStudentForm(instance=student)
     if request.method == 'POST':
@@ -81,8 +88,10 @@ def update_student(request,id):
             return redirect('index')
     else:
         return render(request,'Home/update.html',{'form':form})
-    
+@login_required(login_url='login')   
 def delete_student(request,id):
+    if not request.user.is_superuser:
+        return redirect('index')
     student = Student.objects.get(pk=id)
     student.delete()
     return redirect('index')
